% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum1}
\alias{sum1}
\alias{sum2}
\alias{sum3}
\alias{sum4}
\title{Demonstration functions for calculating the sum of an lvec}
\usage{
sum1(x)

sum2(x)

sum3(x)

sum4(x)
}
\arguments{
\item{x}{an lvec.}
}
\description{
Demonstration functions for calculating the sum of an lvec
}
\details{
The c++ code behind these functions demonstrate three different methods of 
calculating the sum of an lvec object in C++ and one in pure R.
1. Use the get_of_type method of the ldat::vec class to get elements of the
   lvec object as doubles. Sum these. 
2. Use the visitor ldat::lvec_visitor class to construct a visitor that is
   passes to the ldat::vec object. This allows one to write separate methods
   for each type of lvec.
3. Assume that the input is a numeric lvec. Cast the vec to lvec<double> and
   calculate the sum of that. In case one is sure that the lvec is of a
   certain type, this is probably the simplest method.
4. This function is completed completely in R showing how one can process
   data in chunks. If this is possible, this will often give a similar 
   performance as working in C++.
}

